pyproject.toml
[project]
name = "pyros-cli"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.18",
    "art>=6.5",
    "flock-core>=0.4.0b26",
    "loguru>=0.7.3",
    "pillow>=10.4.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.14",
    "questionary>=2.1.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "term-image>=0.7.2",
    "websockets>=15.0.1",
]

[project.scripts]
pyros-cli = "pyros_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.7",
]
